Switching hasUpload, hasRaw and hasForm to the new implementation.
