+++ express-zod-api/src/deep-checks-2.ts
@@ -1,7 +1,12 @@
 import type { $ZodType } from "@zod/core";
 import assert from "node:assert/strict";
 import * as R from "ramda";
-import { z } from "zod";
+import { globalRegistry, z } from "zod";
+import { ezFormBrand } from "./form-schema";
+import { IOSchema } from "./io-schema";
+import { metaSymbol } from "./metadata";
+import { ezUploadBrand } from "./upload-schema";
+import { ezRawBrand } from "./raw-schema";
 
 interface NestedSchemaLookupProps {
   io: "input" | "output";
@@ -23,3 +28,24 @@ export const hasNestedSchema = (
       }) && false,
     () => true,
   )();
+
+export const hasUpload = (subject: IOSchema) =>
+  hasNestedSchema(subject, {
+    condition: (schema) =>
+      globalRegistry.get(schema)?.[metaSymbol]?.brand === ezUploadBrand,
+    io: "input",
+  });
+
+export const hasRaw = (subject: IOSchema) =>
+  hasNestedSchema(subject, {
+    condition: (schema) =>
+      globalRegistry.get(schema)?.[metaSymbol]?.brand === ezRawBrand,
+    io: "input",
+  });
+
+export const hasForm = (subject: IOSchema) =>
+  hasNestedSchema(subject, {
+    condition: (schema) =>
+      globalRegistry.get(schema)?.[metaSymbol]?.brand === ezFormBrand,
+    io: "input",
+  });
